+++ src/VaultSharp/V1/SecretsEngines/PKI/CertificateTidyState.cs
@@ -1,0 +1,17 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace VaultSharp.V1.SecretsEngines.PKI
+{
+    /// <summary>
+    /// Represents the Certificate tidy state.
+    /// </summary>
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum CertificateTidyState
+    {
+        Inactive,
+        Running,
+        Finished,
+        Error
+    }
+}